More Copilot fixes

+++ cmd/zstream/zstream_io.c
@@ -106,7 +106,7 @@ initialize_memory_tracking(void)
 		warnx("unable to read system memory info");
 		payloads.dif_allowed = UINT64_MAX; /* no limit */
 	} else {
-		ssize_t total_mem = pagesize * pages;
+		uint64_t total_mem = pagesize * pages;
 		int64_t flex = (int64_t)total_mem - MEMORY_BASE_CUTOFF;
 		int64_t addl = (double)flex * MEMORY_PCT / 100;
 		payloads.dif_allowed = MEMORY_BASE + MAX(addl, 0);
